+++ docs/changelog.md
@@ -10,6 +10,9 @@
 - Fixed a typo where the CLI setup URL was redirecting to the CI setup
 - Corrected almost all old references to `docs.kurtosistech.com`
 
+### Features
+- Support public ports in Starlark to cover the NEAR usecase
+
 # 0.57.4
 ### Changes
 - Simplified the API by removing the ServiceInfo struct that was adding unnecessary complexity.